example/requirements.txt
django>=4.2
wagtail>=5.0
django-couchbase-orm>=1.0.0
djangorestframework>=3.14
django-oauth-toolkit>=3.0
django-cors-headers>=4.0
gunicorn>=21.0
whitenoise>=6.0
